+++ codec/src/lib.rs
@@ -417,7 +417,7 @@ macro_rules! pdu {
 /// The overall version of the codec.
 /// This must be bumped when backwards incompatible changes
 /// are made to the types and protocol.
-pub const CODEC_VERSION: usize = 34;
+pub const CODEC_VERSION: usize = 35;
 
 // Defines the Pdu enum.
 // Each struct has an explicit identifying number.